Fix error with "compared to"
This is actually an interesting case. Such constructions, referred to as "pseudo-prepositions" were originally annotated in the Penn Treebank, and was also annotated in this EWT example, as a verb (VBN) as PoS but as apearing in a PP, given the pseudo preposition analysis. While we use that analysis in general when there is a nominal complement and both "compared" and "to" are given the dependency "case", the validation doesn't allow "compared" to stay a VERB when the corresponding complement is a clause and these two words would instead have the dependency "mark".

However, on the other hand, LDC's Addendum to the Penn Treebank II Style Bracketing Guidelines, November 2004 on p.38 retracts the classification of "combined with" and "compared with" as pseudo prepositions and says to treat them as verbs whenever possible. This ruling is mostly not followed in EWT, but we nevertheless now follow it for this sentence, and "compared" becomes the "advcl" of the governing clause. This "solves" the issue with the validator.