Fix subordinate verbs, mainly ccomp to parataxis.
This fix is based on examining matches of a pattern in #257 by @nschneider for instances
of ccomp that may not be a ccomp. The false cases of ccomp were changed to either parataxis
or advcl as appropriate.
Several spelling errors that I noticed at the same time have had their lemmas corrected.